+++ src/qwertypy/dsa/strongly_connected_components.py
@@ -1,26 +1,34 @@
-def tarjan(graph):
+def tarjanUndirected(graph):
     low = {}
+    idx  = {}
+    current_id = 0
     visited = set()
     stack = []
     inStack = set()
     
-    def dfs(u):
-        visited.add(u)
-        low[u] = u
-        stack.append(u)
-        inStack.add(u)
+    def dfs(node, parent):
+        nonlocal current_id
+        low[node] = idx[node] = current_id
+        current_id += 1
         
-        for v in graph[u]:
+        visited.add(node)
+        stack.append(node)
+        inStack.add(node)
+        
+        for v in graph[node]:
+            if v == parent: continue
             if v not in visited:
-                dfs(v)
-                low[u] = min(low[u], low[v])
+                dfs(v, node)
+                low[node] = min(low[node], low[v])
             elif v in inStack:
-                low[u] = low[v]
+                low[node] = min(low[node], idx[v])
                 
-        if low[u] == u:
-            while u in inStack:
-                inStack.remove(stack.pop())
+        if low[node] == idx[node]:
+            while node in inStack:
+                v = stack.pop()
+                low[v] = low[node]
+                inStack.remove(v)
     
-    for u in graph:
-        if u not in visited: dfs(u)
+    for u in list(graph.keys()):
+        if u not in visited: dfs(u, None)
     return low