+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0]
+
+### Breaking
+
+-   `qwertypy.dsa.strongly_connected_components.tarjan`
+
+### Added
+
+-   `qwertypy.dsa.strongly_connected_components.tarjanUndirected`
+
 ## [0.6.0]
 
 ### Added